+++ packages/ak-avatar/src/index.js
@@ -1,5 +1,5 @@
 /** @jsx vdom */
-import 'style!./host.less'; // eslint-disable-line import/no-unresolved
+import 'style!./host.less';
 
 import classNames from 'classnames';
 import shadowStyles from './shadow.less';
@@ -14,6 +14,14 @@ const SIZE_ATTRIBUTE_ENUM = {
   invalidDefault: 'medium',
 };
 
+/**
+ * @description The definition for the Avatar component.
+ * @class Avatar
+ * @example @html <ak-avatar src="my/avatar/src/image.png"></ak-avatar>
+ * @example @js import Avatar from 'ak-avatar';
+ * const myAvatar = new Avatar();
+ *
+ */
 const definition = {
   render(elem) {
     const classes = classNames([shadowStyles.locals.img, shadowStyles.locals[elem.size]]);
@@ -30,16 +38,48 @@ const definition = {
     );
   },
   props: {
+    /**
+     * @description The size of the avatar. One of:
+     * 'xsmall', 'small', 'medium', 'large', 'xlarge', 'xxlarge', or 'xxxlarge'.
+     * @memberof Avatar
+     * @instance
+     * @default medium
+     * @type {string}
+     * @example @html <ak-avatar size="large"></ak-avatar>
+     * @example @js avatar.size = 'large';
+     */
     size: enumeration(SIZE_ATTRIBUTE_ENUM)({
       attribute: true,
     }),
+    /**
+     * @description The source URL.
+     * @memberof Avatar
+     * @instance
+     * @type {string}
+     * @example @html <ak-avatar src="my/avatar/src.png"></ak-avatar>
+     * @example @js avatar.src = 'my/avatar/src.png';
+     */
     src: {
       attribute: true,
     },
+    /**
+     * @description The alt text for the Avatar.
+     * @memberof Avatar
+     * @instance
+     * @type {string}
+     * @example @html <ak-avatar alt="Avatar image" src="my/avatar/src.png"></ak-avatar>
+     * @example @js avatar.alt = 'Avatar image';
+     */
     alt: {
       attribute: true,
       default: '',
     },
+    /**
+     * @description The border color for the Avatar.
+     * @memberof Avatar
+     * @instance
+     * @type {string}
+     */
     borderColor: {
       attribute: true,
     },