+++ packages/ak-inline-dialog/CHANGELOG.md
@@ -1,5 +1,41 @@
-<a name="69.2.1"></a>
-## 69.2.1 (2016-07-15)
+<a name="69.4.1"></a>
+## 69.4.1 (2016-07-22)
+
+
+
+<a name="ak-inline-dialog@69.4.1"></a>
+## ak-inline-dialog@69.4.1 (2016-07-22)
+
+
+### Bug Fixes
+
+* fix for the wrong initial position ([18c6bb4](https://bitbucket.org/atlassian/atlaskit/commits/18c6bb4))
+* remove setInterval ([f97ad99](https://bitbucket.org/atlassian/atlaskit/commits/f97ad99))
+
+
+
+<a name="ak-inline-dialog@69.3.0"></a>
+# ak-inline-dialog@69.3.0 (2016-07-22)
+
+
+### Bug Fixes
+
+* fix repositioning in the inline-dialog ([5e0187b](https://bitbucket.org/atlassian/atlaskit/commits/5e0187b))
+
+
+
+<a name="ak-inline-dialog@69.2.2"></a>
+## ak-inline-dialog@69.2.2 (2016-07-18)
+
+
+### Bug Fixes
+
+* **skate:** SkateJS update to beta.10 ([620034a](https://bitbucket.org/atlassian/atlaskit/commits/620034a))
+
+
+### Features
+
+* change 'open' property to boolean for the inline-dialog ([3bd196d](https://bitbucket.org/atlassian/atlaskit/commits/3bd196d))
 
 
 
@@ -25,15 +61,12 @@
 ### Bug Fixes
 
 * remove string attr from target ([26ed13d](https://bitbucket.org/atlassian/atlaskit/commits/26ed13d))
-* remove use of a uniqueWebComponentOld ([200899e](https://bitbucket.org/atlassian/atlaskit/commits/200899e))
 
 
 ### Features
 
 * "doNotMoveInDOM" property (tether fork) ([6405bdc](https://bitbucket.org/atlassian/atlaskit/commits/6405bdc))
 * css options for inline-dialog ([a63837b](https://bitbucket.org/atlassian/atlaskit/commits/a63837b))
-* introducing register.js ([bc60ec1](https://bitbucket.org/atlassian/atlaskit/commits/bc60ec1))
-* return using layer into inline-dialog ([9a905d2](https://bitbucket.org/atlassian/atlaskit/commits/9a905d2))