+++ packages/ak-layer/CHANGELOG.md
@@ -1,5 +1,30 @@
-<a name="56.1.1"></a>
-## 56.1.1 (2016-07-15)
+<a name="56.3.1"></a>
+## 56.3.1 (2016-07-22)
+
+
+
+<a name="ak-layer@56.3.1"></a>
+## ak-layer@56.3.1 (2016-07-22)
+
+
+
+<a name="ak-layer@56.2.0"></a>
+# ak-layer@56.2.0 (2016-07-22)
+
+
+
+<a name="ak-layer@56.1.2"></a>
+## ak-layer@56.1.2 (2016-07-18)
+
+
+### Bug Fixes
+
+* **skate:** SkateJS update to beta.10 ([620034a](https://bitbucket.org/atlassian/atlaskit/commits/620034a))
+
+
+### Features
+
+* change 'open' property to boolean for the inline-dialog ([3bd196d](https://bitbucket.org/atlassian/atlaskit/commits/3bd196d))
 
 
 
@@ -25,7 +50,6 @@
 ### Features
 
 * "doNotMoveInDOM" property (tether fork) ([6405bdc](https://bitbucket.org/atlassian/atlaskit/commits/6405bdc))
-* return using layer into inline-dialog ([9a905d2](https://bitbucket.org/atlassian/atlaskit/commits/9a905d2))