+++ packages/akutil-shared-styles/CHANGELOG.md
@@ -1,5 +1,15 @@
-<a name="33.0.0"></a>
-# 33.0.0 (2016-07-15)
+<a name="33.0.2"></a>
+## 33.0.2 (2016-07-22)
+
+
+
+<a name="akutil-shared-styles@33.0.1"></a>
+## akutil-shared-styles@33.0.1 (2016-07-22)
+
+
+### Bug Fixes
+
+* **package:** Add ak- prefix to editor-*, and fix licensing. ([15751dc](https://bitbucket.org/atlassian/atlaskit/commits/15751dc))